control/pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools_scm>=8", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "RegisterAccessor"
description = "odin-control adapter project for Register Accessor classes, such as ADXDMA or XDMA"
dynamic = ["version"]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: >=3.10",
]
dependencies = [
    "odin-control @ git+https://github.com/odin-detector/odin-control@1.6.0#egg=odin-control"
]
readme = "README.md"

[[project.authors]]
email = "ashley.neaves@stfc.ac.uk"
name = "Ashley Neaves"

[project.scripts]
generate_register_map = "RegisterAccessor.main:main"

[project.urls]
GitHub = "https://github.com/stfc-aeg/memory-accessor"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
# Optional settings for setuptools_scm
root = ".."
write_to = "control/src/RegisterAccessor/_version.py"
version_scheme = "no-guess-dev"
local_scheme = "node-and-date"